Split restore_local_repo.rs into two files, with more descriptive names.
