+++ pageserver/src/lib.rs
@@ -12,16 +12,17 @@ use zenith_metrics::{register_int_gauge_vec, IntGaugeVec};
 pub mod basebackup;
 pub mod branches;
 pub mod http;
+pub mod import_datadir;
 pub mod layered_repository;
 pub mod page_cache;
 pub mod page_service;
 pub mod relish;
 pub mod remote_storage;
 pub mod repository;
-pub mod restore_local_repo;
 pub mod tenant_mgr;
 pub mod tenant_threads;
 pub mod virtual_file;
+pub mod walingest;
 pub mod walreceiver;
 pub mod walrecord;
 pub mod walredo;
